fix: lint and code climate
